Tweaked the spacing of siblings on trains.

+++ src/main/java/mtr/entity/EntityTrain.java
@@ -70,7 +70,7 @@ public abstract class EntityTrain extends EntityCartWorldspikeAdmin implements I
 
 	public abstract int getSiblingSpacing();
 
-	public abstract float getEndSpacing();
+	public abstract int getEndSpacing();
 
 	protected abstract Item getTrainItem();
